Fix GeneralComponents.csproj and refactor AutoLoadItemCacheBase
